refactor: move files from feature userScore
