+++ src/components/tabs/types/typeTabs.ts
